+++ internal/helper/session.go
@@ -1,10 +1,9 @@
-package sessionHelper
+package helper
 
 import (
 	"crypto/subtle"
 	"time"
 
-	"github.com/PurotoApp/authfox/internal/security"
 	"github.com/go-redis/redis"
 )
 
@@ -18,7 +17,7 @@ type sessionPair struct {
 // TODO: use string pointer for UID
 func CreateSession(userID *string, redisVerify, redisSession *redis.Client, verify bool) (sessionPair, error) {
 	// session token
-	token, err := security.RandomString(512)
+	token, err := RandomString(512)
 	if err != nil {
 		return sessionPair{}, err
 	}
