+++ main.go
@@ -2,18 +2,17 @@ package main
 
 import (
 	"github.com/PurotoApp/authfox/internal/endpoints"
-	"github.com/PurotoApp/authfox/internal/gormHelper"
-	"github.com/PurotoApp/authfox/internal/redishelper"
+	"github.com/PurotoApp/authfox/internal/helper"
 	"github.com/PurotoApp/libpuroto/ginHelper"
 	"github.com/gin-gonic/gin"
 )
 
 func main() {
 	// connect to the PostgreSQL
-	pg_conn := gormHelper.ConnectDB()
+	pg_conn := helper.ConnectDB()
 	// Connect to Redis
-	redisVerify := redishelper.Connect(1)
-	redisSession := redishelper.Connect(2)
+	redisVerify := helper.Connect(1)
+	redisSession := helper.Connect(2)
 
 	// migrate all tables
 	endpoints.AutoMigrateAuthfox(pg_conn)
